Another new board layout. Few tidy ups.
Removed the single step circuit, not worth the board space and headache.
Added legend for the IO slot for easy reference. Still toying where to
put the power LED, annoying me with that resistor. Think I'm happy with
it now. Spaced things out a little to make it easy for manufacture, made
sure no traces are near board edges as much as possible.