Added schematic for a nrc2016 USB upgrade board for 68B50
This could be an upgrade board for owners of an rc2014 and reuse their
existing 68B50. It uses the IOSEL lines and also puts an appropriate
oscillator so the 68B50 isn't overclocked to heck.

Also uses a MCP2200 USB to serial converter that does hardware flow
control and will have available GPIO for remote I/O using the Microchip
library (remote reset?).

I've made it so the 68B50 oscillator is optional and can continue to run
from the CPU clk.